BLE: Move SecurityDB related files into generic sources
